vp9: Fix ubsan failure in denoiser.
Fix the overflow for subtraction between two unsigned integers.
https://build.webmproject.org/jenkins/job/libvpx__unit_tests-ubsan/2185/

+++ vp9/encoder/vp9_denoiser.c
@@ -191,7 +191,7 @@ static VP9_DENOISER_DECISION perform_motion_compensation(
     int increase_denoising, int mi_row, int mi_col, PICK_MODE_CONTEXT *ctx,
     int motion_magnitude, int is_skin, int *zeromv_filter, int consec_zeromv,
     int num_spatial_layers, int width) {
-  int sse_diff = ctx->zeromv_sse - ctx->newmv_sse;
+  int64_t sse_diff = (int64_t)ctx->zeromv_sse - (int64_t)ctx->newmv_sse;
   MV_REFERENCE_FRAME frame;
   MACROBLOCKD *filter_mbd = &mb->e_mbd;
   MODE_INFO *mi = filter_mbd->mi[0];
